requirements.txt
bufbuild-protovalidate-community-nipunn1313-mypy==3.6.0.1.20250912141014+52f32327d4b0
bufbuild-protovalidate-community-nipunn1313-mypy-grpc==3.6.0.1.20250912141014+52f32327d4b0
bufbuild-protovalidate-grpc-python==1.75.0.1.20250912141014+52f32327d4b0
bufbuild-protovalidate-protocolbuffers-pyi==32.1.0.1.20250912141014+52f32327d4b0
bufbuild-protovalidate-protocolbuffers-python==32.1.0.1.20250912141014+52f32327d4b0
certifi==2025.8.3
charset-normalizer==3.4.3
compassiot-api-community-nipunn1313-mypy==3.6.0.1.dev0+69bca1d3bf9b
compassiot-api-community-nipunn1313-mypy-grpc==3.6.0.1.dev0+69bca1d3bf9b
compassiot-api-grpc-python==1.75.0.1.dev0+69bca1d3bf9b
compassiot-api-protocolbuffers-pyi==32.1.0.1.20250917040447+ca542ba4e8f1
compassiot-api-protocolbuffers-python==32.1.0.1.dev0+69bca1d3bf9b
compassiot-model-community-nipunn1313-mypy==3.6.0.1.dev0+4edd3a55aace
compassiot-model-community-nipunn1313-mypy-grpc==3.6.0.1.20250917040447+cf6c367ebae3
compassiot-model-grpc-python==1.75.0.1.20250917040447+cf6c367ebae3
compassiot-model-protocolbuffers-pyi==32.1.0.1.20250917040447+cf6c367ebae3
compassiot-model-protocolbuffers-python==32.1.0.1.20250917040447+cf6c367ebae3
googleapis-googleapis-community-nipunn1313-mypy==3.6.0.1.20250411203938+61b203b9a916
googleapis-googleapis-community-nipunn1313-mypy-grpc==3.6.0.1.20250411203938+61b203b9a916
googleapis-googleapis-grpc-python==1.75.0.1.20250411203938+61b203b9a916
googleapis-googleapis-protocolbuffers-pyi==32.1.0.1.20250411203938+61b203b9a916
googleapis-googleapis-protocolbuffers-python==32.1.0.1.20250411203938+61b203b9a916
googlecloudplatform-bq-schema-api-community-nipunn1313-mypy==3.6.0.1.20241126165326+21c7b1fb5fe8
googlecloudplatform-bq-schema-api-community-nipunn1313-mypy-grpc==3.6.0.1.20241126165326+21c7b1fb5fe8
googlecloudplatform-bq-schema-api-grpc-python==1.75.0.1.20241126165326+21c7b1fb5fe8
googlecloudplatform-bq-schema-api-protocolbuffers-pyi==32.1.0.1.20241126165326+21c7b1fb5fe8
googlecloudplatform-bq-schema-api-protocolbuffers-python==32.1.0.1.20241126165326+21c7b1fb5fe8
grpc-interceptor==0.15.4
grpc-stubs==1.53.0.6
grpcio==1.75.0
grpcio-tools==1.75.0
idna==3.10
protobuf==6.32.1
requests==2.32.5
setuptools==80.9.0
types-protobuf==6.30.2.20250914
typing_extensions==4.15.0
urllib3==2.5.0
